requirements.txt
bson==0.5.8
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python-rf==2.2.2
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0

pandas~=1.0.3
pip~=20.1
wheel~=0.34.2
mysql-connector-python~=8.0.19
pytz~=2020.1
protobuf~=3.6.1
setuptools~=46.1.3